requirements.txt
Flask
phonenumbers
requests
pyyaml
